uk-pmrv-app-web/src/app/tasks/aer/verification-submit/pfc/pfc-verification.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';

import { SharedModule } from '@shared/shared.module';
import { AerModule } from '@tasks/aer/aer.module';
import { mockState } from '@tasks/aer/verification-submit/testing/mock-verification-apply-action';
import { TaskSharedModule } from '@tasks/shared/task-shared-module';
import { CommonTasksStore } from '@tasks/store/common-tasks.store';
import { BasePage } from '@testing';
import { KeycloakService } from 'keycloak-angular';

import { PfcVerificationComponent } from './pfc-verification.component';

describe('PfcVerificationComponent', () => {
  let component: PfcVerificationComponent;
  let fixture: ComponentFixture<PfcVerificationComponent>;
  let page: Page;
  let store: CommonTasksStore;

  class Page extends BasePage<PfcVerificationComponent> {
    get rows() {
      return this.queryAll<HTMLTableRowElement>('govuk-table tr')
        .filter((el) => !el.querySelector('th'))
        .map((el) => Array.from(el.querySelectorAll('td')).map((td) => td.textContent.trim()));
    }
    get heading(): string {
      return this.query<HTMLHeadingElement>('h1').textContent.trim();
    }
  }

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      providers: [KeycloakService],
      imports: [SharedModule, RouterTestingModule, TaskSharedModule, AerModule],
    }).compileComponents();
  });

  beforeEach(() => {
    store = TestBed.inject(CommonTasksStore);
    store.setState(mockState);

    fixture = TestBed.createComponent(PfcVerificationComponent);
    component = fixture.componentInstance;
    page = new Page(fixture);
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should render the review groups', () => {
    expect(page.heading).toEqual('Calculation of perfluorocarbons (PFC) emissions');
    expect(page.rows).toEqual([
      ['the reference Anthracite', 'emission source 1 reference', 'Slope', '11332812', ''],
      ['Total emissions', '', '', '11332812 tCO2e', ''],
    ]);
  });
});
